drm/i915: Support enabling panel self-refresh by default
Certain models of Dell XPS support panel self-refresh even though the
default for the processor generation doesn't enable this feature.

Add a quirk to allow defaulting the feature to be enabled for a set of
white-listed systems. The quirk can be driven off the System DMI or
the PCI vendor and device ids.

Signed-off-by: Punit Agrawal <punitagrawal@gmail.com>